refactor: BigDecimal -> BigInt
現在小数点以下を扱える BigDecimal である必要性が無いため
BigDecimal を使うと、整数が必要なときに切り捨てコードを書かなければならないため

+++ app/adapter/src/main/scala/myapp/adapter/account/BankAccountApplication.scala
@@ -5,17 +5,17 @@ import myapp.utility.AppRequestContext
 import scala.concurrent.Future
 
 trait BankAccountApplication {
-  def fetchBalance(accountNo: AccountNo)(implicit appRequestContext: AppRequestContext): Future[BigDecimal]
+  def fetchBalance(accountNo: AccountNo)(implicit appRequestContext: AppRequestContext): Future[BigInt]
 
   def deposit(
       accountNo: AccountNo,
       transactionId: TransactionId,
       amount: Int,
-  )(implicit appRequestContext: AppRequestContext): Future[BigDecimal]
+  )(implicit appRequestContext: AppRequestContext): Future[BigInt]
 
   def withdraw(
       accountNo: AccountNo,
       transactionId: TransactionId,
       amount: Int,
-  )(implicit appRequestContext: AppRequestContext): Future[BigDecimal]
+  )(implicit appRequestContext: AppRequestContext): Future[BigInt]
 }
